generic dense node
