+++ verified/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "t4-verified"
 version = "0.1.3"
-edition = "2021"
+edition = "2024"
 description = "Verus specifications and proofs for the t4 storage model."
 license = "Apache-2.0"
 repository = "https://github.com/XiangpengHao/t4"
